Refactor: 걷기 기록, 순간 기록수 계산 로직 개선 (#193) (#196)

+++ server/src/main/java/backend/section6mainproject/member/dto/MemberControllerDTO.java
@@ -76,8 +76,8 @@ public class MemberControllerDTO {
         private String introduction;
         private WalkLog.WalkLogPublicSetting defaultWalkLogPublicSetting;
         private String imageUrl;
-        private int totalWalkLog;
-        private int totalWalkLogContent;
+        private long totalWalkLog;
+        private long totalWalkLogContent;
         private LocalDateTime createdAt; // 회원가입일시 공개
     }
 }
